make more pretty classes
